Client/Server - March 18th Updates
I added file sending functionality into the server and client, and I
re-worked a portion of the way the client renders graphics. Surfaces are
now separated into individual categories as stated in an enumeration
under the name SurfaceType. I also added in a method that'll convert a
SurfaceTag to its index in the list of surfaces, which will come into
play once items, npcs, and players come into play.

Oh. Lastly, I left the references to the SFML libs in the client. If you
want to switch to the x64 ones, you'll have to do that yourself.