Map Editor - March 23
I started to work on the map editor for the engine. All editors will be
separate from the engine itself, to avoid a bulk of code in the client
and server. As a result, it allows for more room and feature-oriented
possibilities for the editors to be more pleasant when developing;
especially so with the map editor. I am rather anxious to move the
engine along, but I don't want to write sloppy code in the engine. The
way I see it, the editors aren't a part of the engine. So, for now, the
code in the map editor will leave plenty of room for improvement but
it'll get the job done. Please keep in mind that the editor is not yet
finished, and many features still need to be added in.