Framework Conversion
Instead of making a full-sized system like EO, I decided to just resort
to a base-framework. Expect to do a bit of tinkering before making a
game, but at least there will be no bloatware in your system. There is
very little documentation, and there isn't even a sound system in place,
but those will come at later times. Look in the readme for the project
template download.