Bug Fixes 09/07/2016
Fixed known issues with creating projects with the templates regarding
namespace and class names.

Fixed a bug with the Scene Editor form where formatted code copied to
the clipboard would display the wrong information. The editor used the
currently focused scene object to enter in the top and left position for
the formatted code instead of the scene object specified.

The current project templates have these changes.